Move bookList to the containers
